implement fxml
Add FXML support for modular UI development

Current: Project lacks FXML for JavaFX UI definition.

Why: FXML enables modular, readable UI development.

What: Introduce FXML files, update controllers.

Why this way: FXML aligns with JavaFX best practices.

Other: Lays foundation for more maintainable UI.
