+++ src/main/java/exceptions/InvalidCommandException.java
@@ -2,7 +2,7 @@ package exceptions;
 
 /**
  * An exception indicating the inability to decipher the input string.
- * This exception is specific to the Lulu application and is a subclass of {@link LuluException}.
+ * This exception is specific to the lulu.Lulu application and is a subclass of {@link LuluException}.
  */
 public class InvalidCommandException extends LuluException {
     public InvalidCommandException() {
