+++ src/main/java/exceptions/InvalidDateException.java
@@ -2,7 +2,7 @@ package exceptions;
 
 /**
  * An exception indicating the detection of an invalid date.
- * This exception is specific to the Lulu application and is a subclass of {@link LuluException}.
+ * This exception is specific to the lulu.Lulu application and is a subclass of {@link LuluException}.
  */
 public class InvalidDateException extends LuluException {
     public InvalidDateException() {
