+++ src/main/java/exceptions/InvalidStatusUpdateException.java
@@ -2,7 +2,7 @@ package exceptions;
 
 /**
  * An exception indicating an invalid attempt to update the status of a task to a value that has already been set.
- * This exception is specific to the Lulu application and is a subclass of {@link LuluException}.
+ * This exception is specific to the lulu.Lulu application and is a subclass of {@link LuluException}.
  */
 public class InvalidStatusUpdateException extends LuluException {
     public InvalidStatusUpdateException() {
