+++ src/main/java/lulu/Launcher.java
@@ -1,10 +1,13 @@
+package lulu;
+
 import javafx.application.Application;
+import javafxgui.Main;
 
 /**
  * A launcher class to workaround classpath issues.
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Lulu.class, args);
+        Application.launch(Main.class, args);
     }
 }
