+++ src/main/resources/data/lulu.txt
@@ -2,6 +2,5 @@ todo,todo1,false
 event,event1,2022-03-04,2024-05-06,true
 todo,bye,false
 todo,great day,false
-todo,thats great,false
 todo,test,false
-event,event1,2022-03-04,2024-05-06,false
+todo,today,true
