+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-	Hello! I'm Lulu
+	Hello! I'm lulu.Lulu
 	What can I do for you?
 	Got it. I've added this task:
 		[T][ ] borrow book
